requirements.txt
django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
django-ckeditor==6.5.0
djoser==2.1.0
psycopg2~=2.9.3
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.13.0
drf-yasg==1.21.4
django-filter==22.1